requirements.txt
gspread>=6.0.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
pandas>=2.2.0
python-dotenv>=1.0.0
groq>=0.4.1
pytest>=7.4.3
pytest-cov>=4.1.0
matplotlib>=3.8.0